move constructCommitment and constructReveal to common/
